pyproject.toml
[project]
name = "rag-pipeline-exercise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openrouter>=0.1.0",
    "python-dotenv>=1.2.2",
]
